+++ flat_track.py
@@ -156,7 +156,7 @@ class flat_track():
             chromosome = 'chr{}'.format(chromosome)
 
         grp = self.hdf5_handle.create_group('matrix_{}'.format(chromosome))
-        grp.create_dataset('mat', arr.shape, dtype=numpy.float32, data=arr, chunks=True, compression='lzf')
+        grp.create_dataset('mat', arr.shape, dtype=float, data=arr, chunks=True, compression='lzf')
         config.log.info('Added chrom={} to table'.format(chromosome))
 
         self.chrom_names.append(chromosome)
@@ -439,7 +439,7 @@ class flat_track():
                     if len(a) < loc_span: # This should be a very rare case...
                         config.log.warning('Block miss (short)')
                         num_missing = loc_span - len(a)
-                        ad = numpy.zeros(num_missing, dtype=numpy.float32)
+                        ad = numpy.zeros(num_missing, dtype=float)
                         a = numpy.append(a, ad)
 
                     hist += a
@@ -605,14 +605,14 @@ class flat_track():
 
                 # scale center to 0 1000
                 #print('Before', center.size)
-                center = numpy.array(center, dtype=numpy.float)
+                center = numpy.array(center, dtype=float)
                 if center.size != 1000:
                     f = scipyinterp.interp1d(numpy.arange(center.size), center)
                     center = f(numpy.linspace(0, center.size-1, 1000))
                     #print('After', center.size)
 
-                left_flank = numpy.array(left_flank, dtype=numpy.float)
-                rite_flank = numpy.array(rite_flank, dtype=numpy.float)
+                left_flank = numpy.array(left_flank, dtype=float)
+                rite_flank = numpy.array(rite_flank, dtype=float)
 
                 if respect_strand:
                     # positive strand is always correct, so I leave as is.
@@ -932,7 +932,7 @@ class flat_track():
                 cached_chrom = self.get_array_chromosome(l['chr'], read_extend=read_extend) # Will hit the DB if not already in cache
                 # if we are a flat_track, we need to put it to a numpy array:
                 if isinstance(cached_chrom, list):
-                    cached_chrom = numpy.array(cached_chrom, dtype=numpy.float32)
+                    cached_chrom = numpy.array(cached_chrom, dtype=float)
 
             actual_width = cached_chrom.shape[0] - l['left']
 
@@ -963,7 +963,7 @@ class flat_track():
                     a = a[::-1]
             if number_of_tags_in_library:
                 #print(a, number_of_tags_in_library)
-                a = numpy.array(a, dtype=numpy.float)
+                a = numpy.array(a, dtype=float)
                 a /= float(number_of_tags_in_library) # This is 1.0 if norm_by_read_count == False
 
             # bin the data
@@ -1083,7 +1083,7 @@ class flat_track():
             # The interpolation size will be based on
 
             interp_size = distance * scaled_view_fraction
-            center = numpy.array(center, dtype=numpy.float)
+            center = numpy.array(center, dtype=float)
             if center.size != 1000:
                 f = scipyinterp.interp1d(numpy.arange(center.size), center)
                 center = f(numpy.linspace(0, center.size-1, 1000))
